Crud parcial de User

+++ BackEnd/reciclame-backend/src/main/java/com/s1316tjavanext/reciclamebackend/entity/Province.java
@@ -1,10 +1,7 @@
 package com.s1316tjavanext.reciclamebackend.entity;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Data;
 
 @Entity
@@ -14,10 +11,11 @@ import lombok.Data;
 public class Province {
     @Id
     @Schema(description = "Id de la provincia", example = "1")
-    private int id;
+    private Byte id;
 
     @Column(name = "nombre")
     @Schema(description = "Nombre de la provincia", example = "Santa Fe")
     private String name;
+
 
 }
